+++ go/problems/021.go
@@ -11,8 +11,8 @@ import (
 )
 
 func Problem021() string {
-	sum := int64(0)
-	for i := int64(2); i < 10000; i++ {
+	sum := 0
+	for i := 2; i < 10000; i++ {
 		sd := primes.SumDivisors(i)
 		if sd > i && primes.SumDivisors(sd) == i {
 			sum += i + sd
